Rearrange files into a layergroup module
