+++ napari/layers/__init__.py
@@ -13,6 +13,7 @@ from .points import Points
 from .shapes import Shapes
 from .surface import Surface
 from .vectors import Vectors
+from .layergroup import LayerGroup
 
 NAMES = {subclass.__name__.lower() for subclass in all_subclasses(Layer)}
 del all_subclasses
