+++ napari/layers/layergroup/__init__.py
@@ -1,0 +1,1 @@
+from .layergroup import LayerGroup
